+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Unreleased
+- Add enableIpv6 option to clusters
+  [#695](https://github.com/pulumi/pulumi-eks/pull/695)
+  This change upgrades amazon-vpc-cni-k8s to v1.11.
 
 ## 0.39.0 (Released May 9, 2022)
 - Add support for Pulumi AWS 5.x